requirements.txt
python-telegram-bot==20.7
openai==1.12.0
supabase==2.3.1
langgraph==0.0.15
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.0
numpy==1.26.4
tiktoken==0.6.0
pydantic==2.6.1
asyncio==3.4.3
aiohttp==3.9.3
langchain==0.1.9
langchain-openai==0.0.8
fastapi==0.109.2
uvicorn==0.27.1
google-api-python-client==2.118.0
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
jinja2==3.1.3